Create a
Question 1: There should be 2 parentheses after temperature on the first line
Question 4: You can't see options A and B until after you put in your input
Question 5: The variable x isn't properly defined. Also, "==" is used to give x  the value of 4, when they should have just used "x = 4"
Question 6: One line 1, the int function should go in front of input, to convert the input into an integer.  On line 2, there should be 2 equal signs instead of just 1, or else the if statement is giving x a value of 3. Also, there should be a colon after the if statement
Question 7: The if statement should have 2 equal signs instead of just 1, or else it is assigning the variable a value of "Beaker". Also, the if statement is checking the variable "a", when the variable should instead be "answer". There also shouldn't be an indent in front of the else statement, and the else statement should have a colon at the end
Question 8: The code prints "That's good to hear" no matter what the input is
Question 9: I predict that y is false, and that z is true, meaning it will print "Buzz".
I was correct, as the code printed:
x = 5
y = False
z = True
Buzz
Question 10: I guess that the answers will be T, F, T, F, F, T, F, F, T, F, T
My guess was correct
Question 11: I guess the answers will be T, F, T, T, T, T, F, T, F
The actual answers were T, F, T, T, T, F, F, T, and Error
Question 12:  The 3 occupations don't have variables assigned to them.